# docs/openapi/cluster-gateway.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "gcoHmac": {
        "description": "Request-bound HMAC envelope minted by the API Gateway proxy Lambdas (`x-gco-signature-version`, `x-gco-timestamp`, `x-gco-nonce`, `x-gco-signature`): the signature covers the method, exact path and query, body digest, timestamp and nonce under the shared signing key from Secrets Manager. Verified by `gco/services/auth_middleware.py`; a failure is a 403. Original caller identity is authenticated upstream by API Gateway IAM.",
        "in": "header",
        "name": "x-gco-signature",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "The HTTP surface of the internal ALB that fronts a GCO region: the Kubernetes Gateway `gco-system/gco-gateway` (GatewayClass `gco-aws-alb`, implemented by `gateway.k8s.aws/alb`, the AWS Load Balancer Controller) with one HTTPS:443 listener. Its HTTPRoute `gco-routes` sends each request to a Service by longest path prefix; every operation below is one a backend Service serves *and* the route delivers to it, tagged with that Service and annotated with the winning rule (`x-gco-route`). The ALB is `internal` and answers only the API Gateway proxy Lambdas (over Global Accelerator or from the regional bridge's VPC), which sign every request with the HMAC envelope (`x-gco-signature` and companions) that the services verify on every path except `/api/v1/health`, `/healthz`, `/metrics`, `/readyz`. The ALB terminates TLS with a deployment-local certificate and re-encrypts to a TLS proxy sidecar on each pod; the service documents describe the same operations as their applications see them.",
    "title": "GCO Cluster Gateway (internal ALB)",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/inference/{endpoint_name}": {
      "get": {
        "description": "Proxy an endpoint-root request after platform authentication.",
        "operationId": "proxy_inference_root_inference__endpoint_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "endpoint_name",
            "required": true,
            "schema": {
              "title": "Endpoint Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "gcoHmac": []
          }
        ],
        "summary": "Proxy Inference Root",
        "tags": [
          "inference-proxy",
          "Inference"
        ],
        "x-gco-route": {
          "namespace": "gco-system",
          "port": 443,
          "rule": "/inference",
          "service": "inference-proxy"
        }
      },
      "head": {
        "description": "Proxy an endpoint-root request after platform authentication.",
        "operationId": "proxy_inference_root_inference__endpoint_name__head",
        "parameters": [
          {
            "in": "path",
            "name": "endpoint_name",
            "required": true,
            "schema": {
              "title": "Endpoint Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "gcoHmac": []
          }
        ],
        "summary": "Proxy Inference Root",
        "tags": [
          "inference-proxy",
          "Inference"
        ],
        "x-gco-route": {
          "namespace": "gco-system",
          "port": 443,
          "rule": "/inference",
          "service": "inference-proxy"
        }
      },
      "post": {
        "description": "Proxy an endpoint-root request after platform authentication.",
        "operationId": "proxy_inference_root_inference__endpoint_name__post",
        "parameters": [
          {
            "in": "path",
            "name": "endpoint_name",
            "required": true,
            "schema": {
              "title": "Endpoint Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "gcoHmac": []
          }
        ],
        "summary": "Proxy Inference Root",
        "tags": [
          "inference-proxy",
          "Inference"
        ],
        "x-gco-route": {
          "namespace": "gco-system",
          "port": 443,
          "rule": "/inference",
          "service": "inference-proxy"
        }
      }
    },
    "/inference/{endpoint_name}/{upstream_path}": {
      "get": {
        "description": "Proxy an endpoint sub-path after platform authentication.",
        "operationId": "proxy_inference_path_inference__endpoint_name___upstream_path__get",
        "parameters": [
          {
            "in": "path",
            "name": "endpoint_name",
            "required": true,
            "schema": {
              "title": "Endpoint Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "upstream_path",
            "required": true,
            "schema": {
              "title": "Upstream Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "gcoHmac": []
          }
        ],
        "summary": "Proxy Inference Path",
        "tags": [
          "inference-proxy",
          "Inference"
        ],
        "x-gco-route": {
          "namespace": "gco-system",
          "port": 443,
          "rule": "/inference",
          "service": "inference-proxy"
        }
      },
      "head": {
        "description": "Proxy an endpoint sub-path after platform authentication.",
        "operationId": "proxy_inference_path_inference__endpoint_name___upstream_path__head",
        "parameters": [
          {
            "in": "path",
            "name": "endpoint_name",
            "required": true,
            "schema": {
              "title": "Endpoint Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "upstream_path",
            "required": true,
            "schema": {
              "title": "Upstream Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "gcoHmac": []
          }
        ],
        "summary": "Proxy Inference Path",
        "tags": [
          "inference-proxy",
          "Inference"
        ],
        "x-gco-route": {
          "namespace": "gco-system",
          "port": 443,
          "rule": "/inference",
          "service": "inference-proxy"
        }
      },
      "post": {
        "description": "Proxy an endpoint sub-path after platform authentication.",
        "operationId": "proxy_inference_path_inference__endpoint_name___upstream_path__post",
        "parameters": [
          {
            "in": "path",
            "name": "endpoint_name",
            "required": true,
            "schema": {
              "title": "Endpoint Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "upstream_path",
            "required": true,
            "schema": {
              "title": "Upstream Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "gcoHmac": []
          }
        ],
        "summary": "Proxy Inference Path",
        "tags": [
          "inference-proxy",
          "Inference"
        ],
        "x-gco-route": {
          "namespace": "gco-system",
          "port": 443,
          "rule": "/inference",
          "service": "inference-proxy"
        }
      }
    },
...
  },
  "servers": [
    {
      "description": "The region's internal ALB, registered in SSM as `/<project>/alb-hostname-<region>` by ga-registration and kept current by the health monitor; reachable from inside the VPC and from Global Accelerator only. Clients present SNI `backend.<project>.gco.internal`.",
      "url": "https://{alb_hostname}",
      "variables": {
...
    }
...
}